meta: Added clang-format for consistent formatting
I mostly guessed the config based on the current coding style, feel free
to update it if it does't feel good to use.

Having a .clang-format allows people to use the following command:

```
clang-format -i src/*.{c,h}
```

to format the code. That way, multiple people can contribute while
keeping the code in the same style, and making sure everything is
formatted properly.

There are also plugins for most IDE to integrate clang-format and make
it easy to format the code as you write it.